// src/app/modules/course/course.interface.ts
export type ICourseCreateData = {
  title: string;
  code: string;
  credits: number;
  preRequisiteCourses?: IPrerequisiteCourseRequest[];
};

export type IPrerequisiteCourseRequest = {
  courseId: string;
  isDeleted?: null;
};

export type ICourseFilterRequest = {
  searchTerm?: string | undefined;
};